Move user input from index.js to src/user-input.js
refactor
